[PyTest] Adding more test cases(various target dates) / [DateUtil] move date-related functions to date_utils.py, rename dates.py to date_utils.py
